requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.3.0
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyaml==17.12.1
Pygments==2.2.0
PyYAML==3.12
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
